move slip10 into tests and add case.
